Update index.vue
This change closes https://github.com/ClarkLaboratory/IsoVis/issues/2  The issue is that IsoVis is querying the mygene.info API to map gene ids to gene names.  By default, it returns a max of 10 results for each gene matching the query: ("TP53*").  In the case of TP53, there are dozens of genes beginning with the string "TP53" (e.g. TP53I3, TP53BP1, TP53INP2, etc).  It so happens that TP53 is not in the top 10 results returned, and so can never be accessed.  

Adding the "size" parameter with a larger number (we'll try 30 for now) returns more results, including the expected hit for TP53.